% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def-valores.R
\name{def_valores}
\alias{def_valores}
\title{Deflaciona valores nominais em reais}
\usage{
def_valores(.data, date_format, nominal_dates, real_date, index, ...)
}
\arguments{
\item{.data}{Tibble ou data.frame contendo as informações necessárias
para o deflacionamento dos valores.}

\item{date_format}{Formato da data inserida pelo usuário.
Pode ser "ano" ou "dd/mm/yyyy".}

\item{nominal_dates}{Nome da coluna que contém as datas nominais.}

\item{real_date}{String com a data na qual o usuário deseja deflacionar
os valores.}

\item{index}{Índice de preços utilizado para deflacionar os valores.
Pode ser \code{ipca}, \code{igpm}, \code{igpdi}, \code{ipc} ou \code{inpc}.}

\item{...}{Vetor com o nome das colunas que contêm os valores a
serem deflacionados.}
}
\description{
Deflaciona valores nominais em valores reais com base em diversos
índices de preços ao consumidor. Esta função envelopa a função \code{deflate()}
do pacote \code{deflateBR}.
}
