% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate-values.R
\name{deflate_values}
\alias{deflate_values}
\title{Deflate nominal values to real values}
\usage{
deflate_values(.data, date_format, nominal_dates, real_date, index, ...)
}
\arguments{
\item{.data}{Tibble or data.frame containing the necessary information to
deflate the values.}

\item{date_format}{Format of the date inserted by the user. It can be "year"
or "dd/mm/yyyy".}

\item{nominal_dates}{Name of the column containing the nominal dates.}

\item{real_date}{String with the date in which the user wants to deflate the
values.}

\item{index}{Price index used to deflate the values. It can be \code{ipca}, \code{igpm},
\code{igpdi}, \code{ipc} or \code{inpc}.}

\item{...}{Name of the column ou a vector containing the names of the columns
that user want to deflate.}
}
\description{
Deflate nominal values to real values based on several consumer price indices.
This function wraps the \code{deflate()} function from the \code{deflateBR} package,
using \emph{Non Standard Evaluation} programming pattern provided by
\code{rlang} package.
}
